+++ EssentialFeed/EssentialFeed/FeedCache/Infrastructure/CoreData/ManagedCache.swift
@@ -1,0 +1,29 @@
+//
+//  ManagedCache.swift
+//  EssentialFeed
+//
+//  Created by Genrikh Beraylik on 16.12.2021.
+//
+
+import CoreData
+
+@objc(ManagedCache)
+class ManagedCache: NSManagedObject {
+    @NSManaged var timestamp: Date
+    @NSManaged var feed: NSOrderedSet
+    
+    static func find(in context: NSManagedObjectContext) throws -> ManagedCache? {
+        let request = NSFetchRequest<ManagedCache>(entityName: ManagedCache.entity().name!)
+        request.returnsObjectsAsFaults = false
+        return try context.fetch(request).first
+    }
+    
+    static func newUniqueInstance(in context: NSManagedObjectContext) throws -> ManagedCache {
+        try find(in: context).map(context.delete)
+        return ManagedCache(context: context)
+    }
+    
+    var localFeed: [LocalFeedImage] {
+        return feed.compactMap({ ($0 as? ManagedFeedImage)?.local })
+    }
+}
